Cleaned mvp_shell folder and copy-pasted the needed files into mvp_boilerplate
