--- .gitignore
@@ -1,4 +1,0 @@
-raw_data/
-data/.ipynb_checkpoints/
-.DS_Store
-mvp_shell/frontend_repo/.DS_Store
