--- mvp_shell/backend_repo/api_folder/api_file.py
@@ -1,20 +1,0 @@
-from fastapi import FastAPI
-
-api = FastAPI()
-
-# define a root `/` endpoint
-@api.get("/")
-def index():
-    return {"ok": "API connected"}
-
-
-@api.get("/predict")
-def predict(feature1, feature2):
-
-    # model = picle.load_model()
-    # prediction = model.predict(feature1, feature2)
-
-    # Here, I'm only returning the features, since I don't actually have a model.
-    # In a real life setting, you would return the predictions.
-
-    return {'prediction': int(feature1)*int(feature2)}
