--- mvp_shell/frontend_repo/MANIFEST.in
@@ -1,1 +1,0 @@
-include requirements.txt
