--- mvp_shell/frontend_repo/Makefile
@@ -1,2 +1,0 @@
-run_streamlit:
-	streamlit run app.py
