--- mvp_shell/frontend_repo/Procfile
@@ -1,1 +1,0 @@
-web: sh setup.sh && streamlit run app.py
