--- mvp_shell/frontend_repo/setup.sh
@@ -1,13 +1,0 @@
-mkdir -p ~/.streamlit/
-
-echo "\
-[general]\n\
-email = \"benjauzanneau@gmail.com\"\n\
-" > ~/.streamlit/credentials.toml
-
-echo "\
-[server]\n\
-headless = true\n\
-enableCORS = false\n\
-port = $PORT\n\
-" > ~/.streamlit/config.toml
