--- requirements.txt
@@ -1,9 +1,0 @@
-streamlit
-pandas
-numpy
-requests
-altair<5
-watchdog
-jupyter
-fastapi
-uvicorn
